Fix MedicalApplicationProfile ambiguity and module import issues
Critical fixes for GCC 15 module compilation:

1. Fixed MedicalApplicationProfile ambiguity
   - DeviceObjectManager now properly imports mdux.vulkansc.memory module
   - Removed forward declaration, using proper type import instead
   - Fixes "reference to 'MedicalApplicationProfile' is ambiguous" errors

2. Fixed double-promotion warning in MemoryPoolCalculator
   - Explicitly cast safetyMarginMultiplier (float) to double
   - Prevents implicit float-to-double promotion warning
   - Location: MemoryPoolManager.cpp:412

Modified files:
- include/mdux/vulkansc/DeviceObjectManager.cppm
  * Added: import mdux.vulkansc.memory
  * Removed: forward declaration with 'struct' keyword
- src/vulkansc/MemoryPoolManager.cpp
  * Fixed: double-promotion in applySafetyMargin()

Core library now builds successfully on both MSVC and GCC 15.
Remaining warnings in tests/examples will be addressed separately.

+++ src/vulkansc/MemoryPoolManager.cpp
@@ -409,7 +409,7 @@ VkDeviceSize MemoryPoolCalculator::applySafetyMargin(
     VkDeviceSize baseSize,
     const MedicalApplicationProfile& profile) noexcept
 {
-    return static_cast<VkDeviceSize>(static_cast<double>(baseSize) * profile.safetyMarginMultiplier);
+    return static_cast<VkDeviceSize>(static_cast<double>(baseSize) * static_cast<double>(profile.safetyMarginMultiplier));
 }
 
 } // namespace mdux::vulkansc
